Rename quest-badges to badges-list
