+++ frontend/components/badges-list/index.js
