+++ frontend/components/badges-list/styles.css
@@ -1,5 +1,5 @@
 .container {
   border: 1px solid black;
   display: flex;
-  justify-content: space-around;
+  justify-content: space-between;
 }
